AdvanS84/AdvanS8_Standalone_20250604_081543/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
plotly>=5.15.0
requests>=2.31.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
polygon-api-client>=1.12.0
schedule>=1.2.0
pytz>=2023.3
twilio>=8.0.0
trafilatura>=1.6.0
optuna>=3.3.0
xgboost>=1.7.0
tensorflow>=2.13.0
torch>=2.0.0
transformers>=4.30.0